#102 move colored drip particles to shared core
