+++ src/main/resources/assets/growthcraft/particles/colored_drip.json
