+++ src/main/resources/assets/growthcraft/particles/colored_drip_land.json
